Jmcm

+++ app/src/main/java/com/example/jmcm_androidweatherapp/apis/weatherData/main.java
@@ -6,7 +6,7 @@ public class main {
     public double humidity;
     public double temp_min;
     public double temp_max;
-    public int pressure;
+    public double pressure;
 
 }
